+++ packages/components/index.ts
@@ -1,2 +1,3 @@
 export * from './modules/social'
 export * from './modules/navigations'
+export * from './modules/notifications'