+++ packages/components/modules/notifications/NotificationsList/NotificationItem/types.ts
@@ -1,0 +1,12 @@
+import {
+  NotificationItemFragment$data,
+  NotificationItemFragment$key,
+} from '../../../../__generated__/NotificationItemFragment.graphql'
+
+export interface NotificationItemProps {
+  notification: NotificationItemFragment$key
+}
+
+export interface GenericItemProps {
+  notification: NotificationItemFragment$data
+}