+++ packages/components/modules/notifications/constants.ts
@@ -1,0 +1,6 @@
+export const NOTIFICATION_VERB = {
+  commentCreated: 'COMMENT_CREATED',
+  commentReplyCreated: 'COMMENT_REPLY_CREATED',
+}
+
+export const USER_NOTIFICATIONS_KEY = 'user_notifications'