+++ packages/components/modules/notifications/graphql/queries/NotificationsPopover.ts
@@ -1,0 +1,9 @@
+import { graphql } from 'react-relay'
+
+export const NotificationsPopoverQuery = graphql`
+  query NotificationsPopoverQuery {
+    me {
+      ...NotificationUserMenuFragment
+    }
+  }
+`