+++ packages/components/modules/notifications/index.ts
@@ -1,0 +1,5 @@
+export { default as NotificationsPopover } from './NotificationsPopover'
+export type * from './NotificationsPopover/types'
+
+export { default as NotificationList } from './NotificationsList'
+export type * from './NotificationsList/types'